chore/refactor-code-structure: Refactor aicommit and related classes
This commit refactors the aicommit application by removing the AiClient class and consolidating its functionality into new modules for commit and PR description generation. Additionally, environment variable handling has been improved by moving to a common structure, and the ignore file checking mechanism is updated. The PR description generator has been updated to provide more structured output.
