+++ lib/aicommit.rb
@@ -1,56 +1,39 @@
-require "openai"
-require "json"
-require_relative "envs/base"
-require_relative "envs/openai_api_key"
-require_relative "envs/selected_model"
-require_relative "envs/base_branch"
-require_relative "ai_client"
-require_relative "git_client"
+require_relative "commit/generator"
+require_relative "pr/generator"
+require_relative "common/utils/ignore_file_checker"
+require_relative "common/envs/openai_api_key"
+require_relative "common/envs/selected_model"
+require_relative "common/envs/base_branch"
+require_relative "aicommit/version"
 
-class Aicommit
-  def initialize
-    @git_client = GitClient.new
-    @ai_client = AiClient.new
-  end
+module Aicommit
+  # Main entry point for the aicommit gem
 
+  # Generate commit message
   def self.run
-    new.run
+    Commit::Generator.run
   end
 
-  def run
-    staged_changes = git_client.staged_changes
-    commit_message = ai_client.get_commit_message(staged_changes)
-
-    loop do
-      puts "Do you want to keep this commit_message? (Y/R/N) (or Q to quit)"
-      puts ""
-      puts "Commit subject: #{commit_message["subject"]}"
-      puts "Description: #{commit_message["description"]}"
-      puts ""
-      case gets.chomp
-      when /^[Yy]$/
-        git_client.commit_all(commit_message)
-        puts "All changes have been successfully committed."
-        exit
-      when /^[Rr]$/
-        puts "Regenerating..."
-        puts ""
-        commit_message = ai_client.get_commit_message(staged_changes)
-      when /^[Nn]$/
-        puts "Please enter your new commit_message:"
-        commit_message = gets.chomp
-        puts ""
-      when /^[Qq]$/
-        puts "Quit without committing."
-        exit
-      else
-        puts "Invalid command. Please enter Y, N, or Q.".underline
-        puts ""
-      end
-    end
+  # Generate PR description
+  def self.generate_pr_description(base_ref = nil, options = {})
+    PR::Generator.new(base_ref, options).run
   end
 
-  private
+  # Environment utilities
+  def self.set_openai_api_key
+    Common::Envs::OpenaiApiKey.new.update!
+  end
 
-  attr_reader :git_client, :ai_client
+  def self.set_selected_model
+    Common::Envs::SelectedModel.new.update!
+  end
+
+  def self.set_base_branch
+    Common::Envs::BaseBranch.new.update!
+  end
+
+  # Ensure .env is properly set up in gitignore
+  def self.ensure_env_in_ignore_file
+    Common::Utils::IgnoreFileChecker.new.ensure_env_in_ignore_file
+  end
 end
