+++ lib/aicommit/version.rb
@@ -1,3 +1,3 @@
-class Aicommit
+module Aicommit
   VERSION = "1.1.0"
 end
